+++ src/test/java/com/specialpriceshop/raffle/dto/RaffleItemCreateRequestTest.java
@@ -8,6 +8,7 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 import com.specialpriceshop.item.dto.StockCreateRequest;
 import com.specialpriceshop.raffle.domain.Raffle;
 import com.specialpriceshop.raffle.dto.RaffleItemCreateRequest;
+import java.math.BigDecimal;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -20,8 +21,8 @@ class RaffleItemCreateRequestTest {
         final RaffleItemCreateRequest raffleItemCreateRequest = createRaffleItemCreateRequest(
             "상품명",
             "상품설명",
-            5000.0,
-            1000.0,
+            BigDecimal.valueOf(5000),
+            BigDecimal.valueOf(1000.0),
             null
         );
 
@@ -35,15 +36,16 @@ class RaffleItemCreateRequestTest {
     @DisplayName("StockCreateRequest가 null이 아닌 경우")
     void toEntity_stockCreateRequest() {
         final List<StockCreateRequest> stockCreateRequest = List.of(
-            createStockCreateRequest("옵션1", 10L, 10000.0),
-            createStockCreateRequest("옵션2", 10L, 20000.0),
-            createStockCreateRequest("옵션", 10L, 30000.0)
+            createStockCreateRequest("옵션1", 10L, BigDecimal.valueOf(10000.0)),
+            createStockCreateRequest("옵션2", 10L, BigDecimal.valueOf(20000.0)),
+            createStockCreateRequest("옵션", 10L, BigDecimal.valueOf(30000.0))
+
         );
         final RaffleItemCreateRequest raffleItemCreateRequest = createRaffleItemCreateRequest(
             "상품명",
             "상품설명",
-            5000.0,
-            1000.0,
+            BigDecimal.valueOf(5000),
+            BigDecimal.valueOf(1000.0),
             stockCreateRequest
         );
